+++ package.json
@@ -15,13 +15,12 @@
   },
   "dependencies": {
     "gatsby": "^5.12.4",
-    "gatsby-plugin-sass": "^6.14.0",
     "normalize.css": "^8.0.1",
     "package.json": "^2.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "sass": "^1.86.3",
-    "tachyons-sass": "^4.9.5"
+    "tachyons": "^4.12.0"
   },
   "repository": {
     "type": "git",
