+++ src/styles/classes.js
@@ -1,0 +1,35 @@
+// Layout
+export const container = "mw7 center ph4"
+export const section = "pa4 bg-washed-blue br3 shadow-3 mb5"
+export const subsection = "pa4 bg-light-gray br2 shadow-1 mt4 mb4"
+
+// Typography
+export const heading1 = "f1 fw7 blue tc mb4" // larger + vivid
+export const heading3 = "f4 fw6 dark-gray mb2"
+export const paragraph = "lh-copy dark-gray f5 mt3"
+export const list = "list pl3 dark-gray"
+export const link = "link blue hover-underline"
+export const muted = "gray f6 mt2"
+export const bullet = "list-disc list-inside"
+export const centered = "tc"
+
+// Utility
+export const card = "br3 shadow-3 pa4 bg-washed-blue"
+
+// // Layout
+// export const container = "mw7 center ph4"
+// export const section = "pa4 bg-near-white br3 shadow-1 mb4"
+// export const subsection = "pa4 bg-light-gray br2 shadow-1 mt4"
+
+// // Typography
+// export const heading1 = "f2 fw6 dark-blue tc mb4"
+// export const heading3 = "f4 fw6 black-80 mb2"
+// export const paragraph = "lh-copy gray mt3"
+// export const list = "list pl3 gray mb3"
+// export const link = "link blue hover-underline"
+// export const muted = "gray f6 mt2"
+// export const bullet = "list-disc list-inside"
+
+// // Utility
+// export const card = "br3 shadow-1 pa4 bg-washed-blue"
+// export const centered = "tc"
