--- src/styles/common.scss
@@ -1,1 +1,0 @@
-@import "~tachyons-sass/tachyons";
